Moved everything out to the FlameProfiler.py file.
